+++ spec/zip_tricks/rails_streaming_spec.rb
@@ -1,44 +1,113 @@
 require 'spec_helper'
+require 'action_controller'
 
 describe ZipTricks::RailsStreaming do
-  it 'calls the requisite controller methods' do
-    class FakeController
-      include ZipTricks::RailsStreaming
-      attr_reader :response
-      attr_accessor :response_body
-      def initialize
-        @response = Struct.new(:headers, :sending_file).new({})
-      end
-
-      def stream_zip
-        zip_tricks_stream(auto_rename_duplicate_filenames: true) do |z|
-          z.write_deflated_file('hello.txt') do |f|
-            f << 'ßHello from Rails'
-          end
-        end
-      end
-    end
-
-    ctr = FakeController.new
-    ctr.stream_zip
-    response = ctr.response
-    response_body = ctr.response_body
-
-    expect(response.headers['Content-Type']).to eq('application/zip')
-    expect(response.headers['X-Accel-Buffering']).to eq('no')
-    expect(response.sending_file).to be(true)
-
-    ref = StringIO.new('', 'wb')
-    ZipTricks::Streamer.open(ref) do |z|
-      z.write_deflated_file('hello.txt') do |f|
+  module FakeZipGenerator
+    def self.call(streamer)
+      streamer.write_deflated_file('hello.txt') do |f|
         f << 'ßHello from Rails'
       end
     end
 
-    expect(ZipTricks::Streamer).to receive(:new).with(any_args, auto_rename_duplicate_filenames: true).and_call_original
+    def self.generate_reference
+      StringIO.new.binmode.tap do |sio|
+        ZipTricks::Streamer.open(sio) do |streamer|
+          call(streamer)
+        end
+        sio.rewind
+      end
+    end
+  end
 
-    out = StringIO.new('', 'wb')
-    response_body.each.reduce(out, :<<)
-    expect(out.string).to eq(ref.string)
+  class FakeController < ActionController::Base
+    # Make sure both Rack middlewares which are known to cause trouble
+    # are used in this controller, so that we can ensure they get bypassed
+    middleware.use Rack::ETag
+    middleware.use Rack::ContentLength
+
+    include ZipTricks::RailsStreaming
+    def stream_zip
+      zip_tricks_stream(auto_rename_duplicate_filenames: true) do |z|
+        FakeZipGenerator.call(z)
+      end
+    end
+  end
+
+  it 'degrades to a buffered response with HTTP/1.0 and produces a ZIP' do
+    fake_rack_env = {
+      "HTTP_VERSION" => "HTTP/1.0",
+      "REQUEST_METHOD" => "GET",
+      "SCRIPT_NAME" => "",
+      "PATH_INFO" => "/download",
+      "QUERY_STRING" => "",
+      "SERVER_NAME" => "host.example",
+      "rack.input" => StringIO.new,
+    }
+    status, headers, body = FakeController.action(:stream_zip).call(fake_rack_env)
+
+    ref_output_io = FakeZipGenerator.generate_reference
+    out = readback_iterable(body)
+    expect(out.string).to eq(ref_output_io.string)
+
+    expect(status).to eq(200)
+    expect(headers['Content-Type']).to eq('application/zip')
+    expect(headers['ETag']).to be_nil # if the ETag middleware activates it will generate a weak ETag
+    expect(headers['Last-Modified']).to be_kind_of(String)
+    expect(headers['X-Accel-Buffering']).to be_nil # Response gets buffered
+    expect(headers['Transfer-Encoding']).to be_nil
+    # ... and Content-Length may or may not be present
+  end
+
+  it 'uses Transfer-Encoding: chunked with HTTP/1.1 and produces a chunked response' do
+    fake_rack_env = {
+      "HTTP_VERSION" => "HTTP/1.1",
+      "REQUEST_METHOD" => "GET",
+      "SCRIPT_NAME" => "",
+      "PATH_INFO" => "/download",
+      "QUERY_STRING" => "",
+      "SERVER_NAME" => "host.example",
+      "rack.input" => StringIO.new,
+    }
+    status, headers, body = FakeController.action(:stream_zip).call(fake_rack_env)
+
+    ref_output_io = FakeZipGenerator.generate_reference
+    out = decode_chunked_encoding(readback_iterable(body))
+    expect(out.string).to eq(ref_output_io.string)
+
+    expect(status).to eq(200)
+    expect(headers['Content-Type']).to eq('application/zip')
+    expect(headers['ETag']).to be_nil # if the ETag middleware activates it will generate a weak ETag
+    expect(headers['Last-Modified']).to be_kind_of(String)
+    expect(headers['X-Accel-Buffering']).to eq('no')
+    expect(headers['Transfer-Encoding']).to eq('chunked')
+    expect(headers['Content-Length']).to be_nil # Must be unset!
+  end
+
+  def decode_chunked_encoding(io)
+    # Lifted from Net::HTTP mostly
+    StringIO.new.binmode.tap do |dest|
+      begin
+        loop do
+          line = io.readline
+          hexlen = line.slice(/[0-9a-fA-F]+/)
+          len = hexlen.hex
+          break if len == 0 # Terminator (hex 0 followed by \r\n)
+
+          dest.write(io.read(len))
+          io.read(2)   # \r\n
+        end
+      rescue EOFError
+        # nothing
+      ensure
+        dest.rewind
+      end
+    end
+  end
+
+  def readback_iterable(iterable)
+    StringIO.new.binmode.tap do |out|
+      iterable.each { |chunk| out.write(chunk) }
+      out.rewind
+    end
   end
 end